Add French-language weather support :fr:
This addresses snoonetIRC/CloudBot#271

The Weather Underground API supports specifying a return language,
so the translation work locally is mostly swapping a few labels.
While I like to think I know a little French,
I relied entirely on Google Translate for this.

For the sake of completeness though,
I also translated as many error messages as I could.

Also, at the prompting of my IDE, I added type information
to doc-strings where it was missing.

Finally, the global `location_cache` was being kept as a list of
key-value pairs and converted to a dict every time it was checked.
Presumably there was a historical reason for this,
but for now it seems more appropriate to just make it a dict.